refactor: split server.py into separate modules
Move library scanning to library.py, image probing to probing.py,
and cover art source aggregation to sources.py. server.py now only
contains Flask app setup, routes, rate limiting, and startup logic.

Closes #13
